dht data structure: use the correct type for size

+++ lib/dht/data.h
@@ -47,7 +47,7 @@ public:
 	virtual ~Data() {};
 	virtual DataType getDataType() const = 0;
 	virtual bool isEmpty() const = 0;
-	virtual int getSize() const = 0;
+	virtual size_t getSize() const = 0;
 	bool isOld() const;
 	virtual std::string GetStr() const =0;
